Login Version 2 - Integration
Hi Team, 

This new pull request is to integrate my login functionality with the students file upload and download functionality. 

I think we should discuss this today on our daily stand up meeting. It would be good to work with you on this. Our client deadline is Thursday 15th December. We still need time for testing and CMMI maturity reviewing.

Regards,

Mustafa